unet3d_dataset_builder/requirements.txt
numpy>=1.21
h5py>=3.6
tifffile>=2021.11
aicspylibczi>=3.0.3
matplotlib>=3.5
PyYAML>=6.0
